int -> size_t

+++ paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/multiary_infer_sym.cc
@@ -679,7 +679,7 @@ bool BroadcastTensorsOpInferSymbolicShape(
   }
 
   symbol::TensorListShapeOrDataDimExprs out_shapes;
-  for (int i = 0; i < target_rank; i++) {
+  for (size_t i = 0; i < target_rank; i++) {
     out_shapes.emplace_back(out_shape);
   }
